Time: 0 ms (100.00%) | Memory: 8.7 MB (36.33%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -5,7 +5,7 @@ public:
             return false;
         }
         
-        long long rev = 0;
+        long  rev = 0;
         int num = x;
         
         while(num != 0) {
